More tests of calculations to validate figures calculated by mango-explorer.
