+++ test/stoploss.test.ts
@@ -23,7 +23,7 @@ async function testStopLoss() {
   const payer = new Account(
     JSON.parse(
       process.env.KEYPAIR ||
-        fs.readFileSync(os.homedir() + '/.config/solana/devnet.json', 'utf-8'),
+      fs.readFileSync(os.homedir() + '/.config/solana/devnet.json', 'utf-8'),
     ),
   );
   const connection = new Connection(
@@ -156,6 +156,7 @@ async function testStopLoss() {
     0.2276,
     'above',
     51000,
+    false
   );
   console.log('add perp trigger order successful', txid.toString());
   const advanced = await account.loadAdvancedOrders(connection);
