+++ test/testdata/account3/regenerate
