+++ test/testdata/account4/AccountPanel.png
