+++ test/testdata/account4/regenerate
