+++ test/testdata/account5/cache.json
@@ -1,0 +1,11 @@
+{
+    "address": "8mFQbdXsFXt3R3cu3oSNS3bDZRwJRP18vyzd9J278J9z",
+    "executable": false,
+    "lamports": "12082560",
+    "owner": "4skJ85cdxQAFVKbcGgfun8iZPL7BadVYXG3kGEGkufqA",
+    "rent_epoch": "243",
+    "data": [
+        "BwABAAAAAABlGeJYFzcAAAAAAAAAAAAA0NvSYQAAAAB+arx0kxjYuAAAAAAAAAAA0NvSYQAAAACSOgFNhF39DgAAAAAAAAAA0NvSYQAAAACtwJDVrSwAAAAAAAAAAAAA0NvSYQAAAACO5PIf0r8DAAAAAAAAAAAA0NvSYQAAAAAAAAAAAAABAAAAAAAAAAAA0NvSYQAAAADHuriNBgABAAAAAAAAAAAA0NvSYQAAAADKoUW2810BAAAAAAAAAAAA0NvSYQAAAABd3EYDeAMoAAAAAAAAAAAA0NvSYQAAAABqFE+LgxwAAAAAAAAAAAAA0NvSYQAAAAC9e2tToe8AAAAAAAAAAAAA0NvSYQAAAACTh4Va00wFAAAAAAAAAAAA0NvSYQAAAACMpBxNZwYAAAAAAAAAAAAA0NvSYQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACr/bgr9pWKcg8AAAAAAAAAfW/KTOt5fn4PAAAAAAAAANDb0mEAAAAAYZZhSQrgaEIPAAAAAAAAADdFDYM84p1EDwAAAAAAAADQ29JhAAAAAIUYCms10rh3EAAAAAAAAABjeR6HoOFslhAAAAAAAAAA0NvSYQAAAACN6Klk1bUJWg8AAAAAAAAAsOB5MgF2tmQPAAAAAAAAANDb0mEAAAAA2JIjqa0IQEIPAAAAAAAAAEODgb9UekRCDwAAAAAAAADQ29JhAAAAADReISl/WkBCDwAAAAAAAAAsEmw4065AQg8AAAAAAAAA0NvSYQAAAAADJd1j+JAZRA8AAAAAAAAAV7iai6ZVnkQPAAAAAAAAANDb0mEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADssTaSyHEBCDwAAAAAAAAAP2SfBW5xPQg8AAAAAAAAA0NvSYQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAgTGPoWvFwQw8AAAAAAAAAhfEwf8vCB0gPAAAAAAAAANDb0mEAAAAA2UCudSuTs976/////////9lArnUrk7Pe+v/////////Q29JhAAAAAHyXYGjVpjHDAgAAAAAAAAC8QAFoi7CH3wEAAAAAAAAA0NvSYQAAAABQGTn1zyDwfe//////////UBk59c8g8H3v/////////9Db0mEAAAAA13/fY15IeV71/////////5MV9jBit0i88v/////////Q29JhAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAPqbMtP2S19s9P/////////01TffyJFd1/P/////////0NvSYQAAAAANm/JrQVRTu/f/////////DZvya0FUU7v3/////////9Db0mEAAAAABnxCgzEAYu0BAAAAAAAAAAZ8QoMxAGLtAQAAAAAAAADQ29JhAAAAAHJjA10cH88P9/////////9yYwNdHB/PD/f/////////0NvSYQAAAADHIR5z5iUwjf//////////xyEec+YlMI3//////////9Db0mEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADQ29JhAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
+        "base64"
+    ]
+}
