+++ test/testdata/account6/account.json
@@ -1,0 +1,11 @@
+{
+    "address": "DRUZRfLQtki4ZYvRXhi5yGmyqCf6iMfTzxtBpxo6rbHu",
+    "executable": false,
+    "lamports": "30791040",
+    "owner": "4skJ85cdxQAFVKbcGgfun8iZPL7BadVYXG3kGEGkufqA",
+    "rent_epoch": "245",
+    "data": [
+        "AQABAAAAAADKISAtTF7lMdQarmxDZG/wXw1EdsMrRmiMvKrmbhkiIbvxamIzDH990RAEJs6CrXl6CFbOk/3Sn4lB//dHl8AAAQEBAQAAAAAAAAAAAAAABKZdbzEbGNZWAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAKFz7oDbrwAAAAAAAAAAAAAkkCPkm/XLlAgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAJFTatxUxIR3PwAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAcm3Vr0OSCQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAL0egN3L/5kBAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQPV4zkOKDqC5VXaIVyZDONipRkZxR6iTqZVINWYrdlPaW2fB4Qjqgmqq0At4aMHMxqiLO8j3lGfJFby0fChexz3Me+YR0y5SMLI1Z09RiqVuHS2iZg9f6eB9LZw6UMXt7atkJyuygIOl1HBwGg0pc83Ou4UNUvmaY0bGQtu9dMAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAGxVS+THu8Qiz7lAJAQlt3EnqtYIh7R8VssjZDgE8hcRAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAIaj///////85aYI7FPgKEeQEAQAAAAAAIz4pQFiRm/n8/////////yM+KUBYkZv5/P////////9FikIAAAAAANGJQgAAAAAAAAAAAAAAAAAAAAAAAAAAAFyVaTkKAAAAwfwCAAAAAAAgglSuTTYbyKT8FgkAAAAAHQZSyNqWiGoCAAAAAAAAAF2v8seQoN6GAQAAAAAAAADRKwMAAAAAANErAwAAAAAAAAAAAAAAAAAAAAAAAAAAAAYYBAAAAAAA8PL8//////+EghE+tGfxP4DCzgAAAAAAIvetk8djXmDx/////////yL3rZPHY15g8f/////////NEAQAAAAAAM0QBAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA8rj///////+5UBdoedCYvXrDDAAAAAAAkwxJsvtz3hL2/////////0+iX3//4q1w8//////////NEwMAAAAAAM0TAwAAAAAAAAAAAAAAAAAAAAAAAAAAADgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEf3///////8RKKNnLySYkihsAwAAAAAAfDprS7XgDMb1/////////3Z0cFeHJgsx9f////////8HTwsAAAAAAAdPCwAAAAAAAAAAAAAAAAAAAAAAAAAAANvqqXgJAAAAaBAAAAAAAAAASZ0KAABw5/sZ/P//////5rnH8dEV/pH3/////////+a5x/HRFf6R9//////////PkwMAAAAAAM+TAwAAAAAAAAAAAAAAAAAAAAAAAAAAAGstYR1DAAAAonz+//////8gdPerVI8Ii4mMHAAAAAAA4Zoft92lwZL//////////+GaH7fdpcGS///////////dzQ4AAAAAAN3NDgAAAAAAAAAAAAAAAAAAAAAAAAAAAA5xW4ZIAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACDeAYAAAAAAIN4BgAAAAAAAAAAAAAAAAAAAAAAAAAAALGMo2gPAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABwcHBwcHAAAAAAAACAgICAgIDAwMDAwMAgICAgICAQEBAQEBCQkJCQkJAwMDAwMD/////////////////////wABAAEAAQABAAEAAQABAAEAAQABAAEAAQABAAEAAQABAAEAAQABAAEAAQABAAEAAQAAAAAAAAAAAAAAAAAAAADQCd7//////6MuAAAAAAAAMPYhAAAAAAAbLwAAAAAAAM4J3v//////Zy4AAAAAAAAy9iEAAAAAAFcvAAAAAAAAzAne//////8rLgAAAAAAADT2IQAAAAAAky8AAAAAAABLR+D//////5MHAAAAAAAAo7gfAAAAAAABCAAAAAAAAEpH4P//////kwcAAAAAAACluB8AAAAAAAwIAAAAAAAASUfg//////+TBwAAAAAAAKe4HwAAAAAAFggAAAAAAACDx+D//////26TAAAAAAAAfTgfAAAAAADplAAAAAAAAIHH4P//////sJIAAAAAAAB/OB8AAAAAAKeVAAAAAAAAf8fg///////ykQAAAAAAAIE4HwAAAAAAZJYAAAAAAADjf/z//////4JRAAAAAAAAHYADAAAAAAAsYwAAAAAAAOF//P//////GlEAAAAAAAAfgAMAAAAAACxjAAAAAAAA33/8//////+xUAAAAAAAACGAAwAAAAAALGMAAAAAAABt89///////3ARBQAAAAAAkwwgAAAAAAB6HgUAAAAAAGvz3///////3AoFAAAAAACVDCAAAAAAAPAkBQAAAAAAafPf//////9OBAUAAAAAAJcMIAAAAAAAbCsFAAAAAABqBr3//////+B+BgAAAAAAlvlCAAAAAACXjwYAAAAAAGgGvf//////nXYGAAAAAACY+UIAAAAAAAqYBgAAAAAAZga9//////9PbgYAAAAAAJr5QgAAAAAAc6AGAAAAAAAA9+3//////6AjAAAAAAAAAAkSAAAAAAD8IwAAAAAAAP727f//////ciMAAAAAAAACCRIAAAAAACokAAAAAAAA/Pbt//////9EIwAAAAAAAAQJEgAAAAAAVyQAAAAAAAAzqr7//////yI5AAAAAAAAzVVBAAAAAAC1OQAAAAAAADGqvv//////2DgAAAAAAADPVUEAAAAAAP45AAAAAAAAL6q+//////+OOAAAAAAAANFVQQAAAAAARzoAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAgts1L34BAACD2zUvfgEAAITbNS9+AQAAhds1L34BAACG2zUvfgEAAJHbNS9+AQAABeU1L34BAABpITUvfgEAAAblNS9+AQAAbCE1L34BAAAH5TUvfgEAAG4hNS9+AQAAhOA1L34BAACF4DUvfgEAAIbgNS9+AQAAh+A1L34BAACI4DUvfgEAAIngNS9+AQAAD881L34BAAAQzzUvfgEAABHPNS9+AQAAEs81L34BAAATzzUvfgEAABTPNS9+AQAAJtQ1L34BAAAn1DUvfgEAACjUNS9+AQAAKdQ1L34BAAAq1DUvfgEAACvUNS9+AQAACcs1L34BAAAKyzUvfgEAAAvLNS9+AQAADMs1L34BAAANyzUvfgEAAA7LNS9+AQAAtOg1L34BAAC16DUvfgEAALboNS9+AQAAt+g1L34BAAC46DUvfgEAALnoNS9+AQAA1NA1L34BAADV0DUvfgEAANbQNS9+AQAA19A1L34BAADY0DUvfgEAANnQNS9+AQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
+        "base64"
+    ]
+}
