+++ test/testdata/account6/cache.json
@@ -1,0 +1,11 @@
+{
+    "address": "8mFQbdXsFXt3R3cu3oSNS3bDZRwJRP18vyzd9J278J9z",
+    "executable": false,
+    "lamports": "12082560",
+    "owner": "4skJ85cdxQAFVKbcGgfun8iZPL7BadVYXG3kGEGkufqA",
+    "rent_epoch": "245",
+    "data": [
+        "BwABAAAAAAA51sVtNDQAAAAAAAAAAAAAetbWYQAAAAC6SQwCK6cSpwAAAAAAAAAAetbWYQAAAAAD54wo7b0LDQAAAAAAAAAAetbWYQAAAACJsyJqoiUAAAAAAAAAAAAAetbWYQAAAACMKO0NvjADAAAAAAAAAAAAetbWYQAAAAAAAAAAAAABAAAAAAAAAAAAetbWYQAAAAA4RUdy+f8AAAAAAAAAAAAAetbWYQAAAAAW2c73UzMBAAAAAAAAAAAAetbWYQAAAAC6uI0G8PYlAAAAAAAAAAAAetbWYQAAAADg21FrfBcAAAAAAAAAAAAAetbWYQAAAADqZYJDor8AAAAAAAAAAAAAetbWYQAAAACTGARWDp0EAAAAAAAAAAAAetbWYQAAAACo1vBkXwUAAAAAAAAAAAAAetbWYQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAACr/bgr9pWKcg8AAAAAAAAAfW/KTOt5fn4PAAAAAAAAAHzW1mEAAAAAvttQwWyYbUIPAAAAAAAAAFjMse7KwtdEDwAAAAAAAAB81tZhAAAAAB7PdKQvQCabEAAAAAAAAAALpm+fDGBXvRAAAAAAAAAAfNbWYQAAAAAe9g7LkpKaWg8AAAAAAAAA5kD52dq2+GUPAAAAAAAAAHzW1mEAAAAAFssqHa4IQEIPAAAAAAAAANmR0TdBjERCDwAAAAAAAAB81tZhAAAAADReISl/WkBCDwAAAAAAAAAsEmw4065AQg8AAAAAAAAAfNbWYQAAAAAzoyaapZMZRA8AAAAAAAAA8VB1z9AOoUQPAAAAAAAAAHzW1mEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAHGAaS20IkBCDwAAAAAAAAD5L/jri35SQg8AAAAAAAAAfNbWYQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABFwTPIWihxQw8AAAAAAAAAsGCNQusXFEgPAAAAAAAAAHzW1mEAAAAAqRjqbTY671n6/////////6kY6m02Ou9Z+v////////981tZhAAAAAMbbFM+yhJtpAgAAAAAAAAAGhbXOaI7xhQEAAAAAAAAAfNbWYQAAAAALGiWHGZdqee//////////CxolhxmXannv/////////3zW1mEAAAAAOVnTA79l3hL2//////////Xu6dDC1K1w8/////////981tZhAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAMSDJcxswx9s9P////////++vSrYPgke1/P/////////fNbWYQAAAABngWsBSRigu/f/////////Z4FrAUkYoLv3/////////3zW1mEAAAAAKlBZ4XgRAe4BAAAAAAAAACpQWeF4EQHuAQAAAAAAAAB81tZhAAAAACv805m29/X69v////////8r/NOZtvf1+vb/////////fNbWYQAAAADHIR5z5iUwjf//////////xyEec+YlMI3//////////3zW1mEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAB61tZhAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
+        "base64"
+    ]
+}
